// types/extended.ts
export interface UserInfo {
  id: string;
  username: string;
  image?: string | null;
  name?: string | null;
  surname?: string | null;
}
// ...
export type AssignedItemType = "task" | "mindMap";
// ...
export interface WorkspaceRecentActivityAssignedToItem {
  user: {
    id: string;
    image: string | null;
    username: string;
  };
  id: string;
  userId: string;
  mindMapId?: string;
  taskId?: string;
}

export interface WorkspaceRecentActivity {
  id: string;
  title: string;
  emoji: string;
  type: AssignedItemType;
  updated: {
    at: Date;
    by?: UserInfo | null;
  };
  starred: boolean;
  tags: Tag[];
  assignedTo: WorkspaceRecentActivityAssignedToItem[];
  link: string;
}